fix lints

+++ pallas-applying/src/shelley_ma.rs
@@ -408,7 +408,8 @@ fn compute_script_hash(script: &NativeScript) -> PolicyId {
     pallas_crypto::hash::Hasher::<224>::hash(&payload)
 }
 
-// Checks all certificates in order, and counts the relevant ones for computing deposits.
+// Checks all certificates in order, and counts the relevant ones for computing
+// deposits.
 #[allow(clippy::too_many_arguments)]
 fn check_certificates(
     cert_opt: &Option<Vec<Certificate>>,
@@ -591,7 +592,7 @@ fn check_pool_retirement(
     if !ps.pool_params.contains_key(pool_hash) {
         return Err(ShelleyMA(PoolNotRegistered));
     }
-    if (*cepoch < *repoch) & (*repoch <= *cepoch + *emax as u64) {
+    if (*cepoch < *repoch) & (*repoch <= *cepoch + *emax) {
         ps.retiring.insert(*pool_hash, *repoch);
         Ok(())
     } else {
